lambda/requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
boto3==1.42.76
botocore==1.42.76
certifi==2026.2.25
click==8.3.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
jmespath==1.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.24.4
opencv-python-headless==4.11.0.86
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
protobuf==7.34.1
pyclipper==1.4.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rapidocr-onnxruntime==1.2.3
rich==14.3.3
s3transfer==0.16.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
